requirements.txt
customtkinter
requests
psutil
Pillow
pycryptodome
Faker
phonenumbers
